Move to FeedBackgroundWorker
